tests/wpt/tests/html/semantics/forms/the-selectmenu-element/selectmenu-tab-navigation.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<input id="input1">
<selectmenu id="selectmenu">
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>
<input id="input3">

<script>
promise_test(async () => {
  const TAB_KEY = "\uE004";

  const input1 = document.getElementById("input1");
  const selectmenu = document.getElementById("selectmenu");

  input1.focus();
  assert_equals(document.activeElement.id, "input1", "input1 should be active");

  await test_driver.send_keys(input1, TAB_KEY);
  assert_equals(document.activeElement.id, "selectmenu", "selectmenu should be active");

  await test_driver.send_keys(selectmenu, TAB_KEY);
  assert_equals(document.activeElement.id, "input3", "input3 should be active");
}, "Check that <selectmenu> occupies just one slot in tab navigation.");
</script>
